# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MoneyTor Finance Tracker API
  description: Backend API for the finance tracker frontend. All dates/times are ISO 8601 / RFC3339. Authenticated endpoints require Bearer JWT.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1 base path (prepend your host, e.g. https://api.example.com)
security:
  - bearerAuth: []
...
paths:
...
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string, description: JWT refresh token }
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessAuthData' }
        '400':
          description: Bad request
        '401':
          description: Unauthorized or invalid refresh token
  /auth/me:
    get:
      tags: [auth]
      summary: Get current user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/SuccessEnvelope' }
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/User' }
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from login or refresh
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data: { description: Response payload; omit on error }
...
    User:
      type: object
      properties:
        uuid: { type: string, format: uuid }
        email: { type: string }
        name: { type: string, nullable: true }
        role: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SuccessAuthData:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                token: { type: string, description: Access JWT }
                refreshToken: { type: string, description: Refresh JWT }
                user: { $ref: '#/components/schemas/User' }
...